Move as many functions as possible to kOS.Safe
